requirements.txt
requests
speedtest-cli
